+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Indent `fdef` (clojure.spec) like a `def`.
 * The results of `clojure-project-dir` are cached by default to optimize performance.
+* Add `shadow-cljs.edn` to the default list of build tool files.
 
 ## 5.7.0 (2018-04-29)